Rough draft of a bonfire module.
I added all the minimum requirements for the bonfire module, but I still
need to connect some other things together. I need to run some actual
cron tests in the morning, and during that point I'll also have to
create another migration to create a table for cron logs that can be
viewed in the report.