Added settings manager, overhauled exports list
Huge updates here. Added a fully functioning settings managing,
including the ability to activate/deactivate cron jobs, change the
filename of the file that will be exported, and change the memory and
time limits for each job.

Also overhauled the exports list and revamped the functionality
considerably. It now displays all cron jobs once on the main page, along
with the latest file exported in each. You can download a zip of all
files on the page. You can also open up each cron job individually,
export each file in the history of the job, and also download a zip of
all files for one cron job.